Moved non-index EditableDisplay files into a 'helpers' folder.
