+++ app/modules/contentItems/components/EditableDisplay/helpers/TypeBlockWrapper.js
@@ -1,7 +1,6 @@
 // @flow
 
 import * as React from 'react';
-
 import { Icon } from 'semantic-ui-react';
 
 type PassedProps = {
